modified:   curaj_cpp/student_details.cpp

+++ curaj_cpp/student_details.cpp
@@ -8,7 +8,7 @@ class student
     private:
     int roll;
     char name;
-    float marks;
+    int marks;
     public:
     void getdata()
     {
